feat: sprint11 изменения по сущностям, дальше переход на БД.

+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -43,34 +43,34 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User getUserById(@PathVariable(value = "id") int id) {
+    public User getUserById(@PathVariable(value = "id") Long id) {
         log.debug("[getUserById] Start.");
         return userService.getUserById(id);
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
-    public List<User> getCommonFriends(@PathVariable(value = "id") int id,
-                                       @PathVariable(value = "otherId") int otherId) {
+    public List<User> getCommonFriends(@PathVariable(value = "id") Long id,
+                                       @PathVariable(value = "otherId") Long otherId) {
         log.debug("[getCommonFriends] Start.");
         return userService.getCommonFriends(id, otherId);
     }
 
     @PutMapping("/{id}/friends/{friendId}")
-    public void addFriend(@PathVariable(value = "id") int id,
-                          @PathVariable(value = "friendId") int friendId) {
+    public void addFriend(@PathVariable(value = "id") Long id,
+                          @PathVariable(value = "friendId") Long friendId) {
         log.debug("[addFriend] Start.");
         userService.addFriend(id, friendId);
     }
 
     @GetMapping("{id}/friends")
-    public List<User> getFriends(@PathVariable(value = "id") int id) {
+    public List<User> getFriends(@PathVariable(value = "id") Long id) {
         log.debug("[getFriends] Start.");
         return userService.getFriends(id);
     }
 
     @DeleteMapping("/{id}/friends/{friendId}")
-    public void removeFriend(@PathVariable(value = "id") int id,
-                             @PathVariable(value = "friendId") int friendId) {
+    public void removeFriend(@PathVariable(value = "id") Long id,
+                             @PathVariable(value = "friendId") Long friendId) {
         log.debug("[removeFriend] Start.");
         userService.removeFriend(id, friendId);
     }
